pyproject.toml
[tool.poetry]
name = "orienteer"
version = "0.1.0"
description = ""
authors = ["pguin-sudo <mailfordevelopmentandstuff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.31"
disnake = "^2.9.2"
loguru = "^0.7.2"
uvloop = "^0.19.0"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.5"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
